requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.4.0
Django==5.2.5
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-filter==24.2
django-jazzmin==3.0.1
django-rest-knox==5.0.2
django-sass-processor==1.4.1
django-taggit==6.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
filelock==3.18.0
idna==3.7
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
libsass==0.23.0
oauthlib==3.2.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.2
requests-oauthlib==2.0.0
rpds-py==0.27.0
setuptools==80.9.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
Strip==0.0.1.dev14
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.1
virtualenv==20.33.1
wheel==0.45.1
whitenoise==6.6.0






